show multiple parents in same row (#209)
For a multi-value field displayed in Views, the setting for **Multiple field settings** / *Display all values in the same row* has the help text “If checked, multiple values for this field will be shown in the same row. If not checked, each value in this field will create a new row. If using group by, make sure to group by "Entity ID" for this setting to have any effect.”

This change updates that setting for the Parents (`field_member_of`) field so that if a Repository Item is a member of multiple collections it will not be shown twice in the content list, and its parents will be in the same table cell separated by commas.